Replace exposed-java-time with exposed-kotlin-datetime and add Workshops component
This update implements several changes in the frontend and backend of the app including the swap from exposed-java-time to exposed-kotlin-datetime in build.gradle.kts for improved runtime efficiency. In the frontend, a new component named Workshops has been added for managing workshop data. Furthermore, backend repository files WorkshopRegistrationRepository and WorkshopRepository are updated respectively for seamless integration with exposed-kotlin-datetime.